+++ CHANGELOG.md
@@ -43,6 +43,7 @@
   - support for selecting custom chat wallpaper #4306
   - support for themes #4304
 - improve keyboard and screen-reader accessibility #4210
+- make backup and key export work in browser #4303
 
 ## Changed
 - style: avoid scrolling to account list items such that they're at the very edge of the list #4252